+++ src/examples/using-custom-network.ts
@@ -6,9 +6,9 @@ import {
   PermissionScope,
   TezosOperationType,
   OperationResponseOutput,
-  PermissionResponseOutput
+  PermissionResponseOutput,
+  PartialTezosTransactionOperation
 } from '..' // Replace '..' with '@airgap/beacon-sdk'
-import { PartialTezosTransactionOperation } from '../types/tezos/PartialTezosOperation'
 
 const client = new DAppClient({ name: 'My Sample DApp' })
 
@@ -29,8 +29,7 @@ client
     network,
     scopes
   })
-  .then((permissionResponse: PermissionResponseOutput) => {
-    // Check if operation permissions were granted
+  .then(async (permissionResponse: PermissionResponseOutput) => {
     if (
       permissionResponse.scopes.some(
         (permission: PermissionScope) => permission === PermissionScope.OPERATION_REQUEST
@@ -41,15 +40,14 @@ client
         amount: '1234567',
         destination: 'tz1MJx9vhaNRSimcuXPK2rW4fLccQnDAnVKJ'
       }
-      client
-        .requestOperation({ operationDetails: [operation] })
-        .then((operationResponse: OperationResponseOutput) => {
-          console.log(
-            'operation was successfully broadcast to the network with the hash: ',
-            operationResponse
-          )
-        })
-        .catch((operationError: ErrorResponse) => console.error(operationError))
+      const operationResponse: OperationResponseOutput = await client.requestOperation({
+        operationDetails: [operation]
+      })
+
+      console.log(
+        'operation was successfully broadcast to the network with the hash: ',
+        operationResponse.transactionHash
+      )
     }
   })
   .catch((permissionError: ErrorResponse) => console.error(permissionError))