+++ src/types/storage/StorageKeyReturnType.ts
@@ -6,6 +6,7 @@ import {
   AppMetadata,
   PermissionInfo
 } from '../..'
+import { MatrixState } from '../../matrix-client/MatrixClientStore'
 
 export interface StorageKeyReturnType {
   [StorageKey.TRANSPORT_P2P_PEERS]: P2PPairingRequest[]
@@ -15,4 +16,5 @@ export interface StorageKeyReturnType {
   [StorageKey.APP_METADATA_LIST]: AppMetadata[]
   [StorageKey.PERMISSION_LIST]: PermissionInfo[]
   [StorageKey.BEACON_SDK_VERSION]: string | undefined
+  [StorageKey.MATRIX_PRESERVED_STATE]: Partial<MatrixState>
 }